splitting file
